+++ gamedata/shaders/sprites.glsl
@@ -8,12 +8,12 @@ uniform mat4 view_projection_matrix;
 // attributes
 in vec3 in_pos0;
 in vec2 in_texcoord0;
-in float in_color0; // palette index
+in int in_color0; // palette index
 
 // pass to fragment shader
 out vec2 Texcoord;
 out vec3 Position;
-flat out float PaletteIndex;
+flat out int PaletteIndex;
 
 // entry point
 void main() 
@@ -37,7 +37,7 @@ uniform sampler2D tex_3; // palettes table
 // passed from vertex shader
 in vec2 Texcoord;
 in vec3 Position;
-flat in float PaletteIndex;
+flat in int PaletteIndex;
 
 // result
 out vec4 FinalColor;
@@ -53,7 +53,7 @@ void main()
         discard;
 
     // fetch pixel color
-    vec4 pixelColor = texelFetch(tex_3, ivec2(int(pal_color), int(PaletteIndex)), 0);
+    vec4 pixelColor = texelFetch(tex_3, ivec2(int(pal_color), PaletteIndex), 0);
     pixelColor.a = 1.0f;
     FinalColor = clamp(pixelColor, 0.0f, 1.0f);
 }
